fix: move generation of avatar loader settings
